Changes to compile with Xcode 12.2 on Apple Silicon for M1 MacBook Pro
The Apple 13” MacBook Pro (with Apple Silicon - M1) and a Samsung C49RG9x failed to support the full 5120 x 1440 resolution (wouldn’t go past 3840x1080) despite it being supported by a late 2016 touchbar 15” MacBook Pro.

Building this in Xcode 12.2 with the necessary changes to get it compiling with modern macOS SDK and then choosing 5120 x 1440 from the RDM Status item menu was sufficient to get it working.

Adding an Xcode project for ease of building withing Xcode (not required).